+++ mrserver/mrparser/datatime.go
@@ -1,0 +1,33 @@
+package mrparser
+
+import (
+	"net/http"
+	"time"
+
+	"github.com/mondegor/go-webcore/mrctx"
+	"github.com/mondegor/go-webcore/mrserver"
+	"github.com/mondegor/go-webcore/mrserver/mrreq"
+)
+
+type (
+	DateTime struct {
+	}
+)
+
+// Make sure the DateTime conforms with the mrserver.RequestParserDateTime interface
+var _ mrserver.RequestParserDateTime = (*DateTime)(nil)
+
+func NewDateTime() *DateTime {
+	return &DateTime{}
+}
+
+func (p *DateTime) FilterDateTime(r *http.Request, key string) time.Time {
+	value, err := mrreq.ParseDateTime(r, key, false)
+
+	if err != nil {
+		mrctx.Logger(r.Context()).Warn(err)
+		return time.Time{}
+	}
+
+	return value
+}