+++ mrserver/mrparser/string.go
@@ -1,0 +1,50 @@
+package mrparser
+
+import (
+	"net/http"
+
+	"github.com/mondegor/go-webcore/mrctx"
+	"github.com/mondegor/go-webcore/mrserver"
+	"github.com/mondegor/go-webcore/mrserver/mrreq"
+)
+
+type (
+	String struct {
+		pathFunc mrserver.RequestParserParamFunc
+	}
+)
+
+// Make sure the String conforms with the mrserver.RequestParserString interface
+var _ mrserver.RequestParserString = (*String)(nil)
+
+func NewString(pathFunc mrserver.RequestParserParamFunc) *String {
+	return &String{
+		pathFunc: pathFunc,
+	}
+}
+
+func (p *String) PathParamString(r *http.Request, name string) string {
+	return p.pathFunc(r, name)
+}
+
+// RawParamString - returns nil if the param not found
+func (p *String) RawParamString(r *http.Request, key string) *string {
+	if !r.URL.Query().Has(key) {
+		return nil
+	}
+
+	value := r.URL.Query().Get(key)
+
+	return &value
+}
+
+func (p *String) FilterString(r *http.Request, key string) string {
+	value, err := mrreq.ParseStr(r, key, false)
+
+	if err != nil {
+		mrctx.Logger(r.Context()).Warn(err)
+		return ""
+	}
+
+	return value
+}