+++ mrtype/string_pointer.go
@@ -7,3 +7,13 @@ func StringPointer(value string) *string {
 
 	return &value
 }
+
+func StringPointerCopy(value *string) *string {
+	if value == nil || *value == "" {
+		return nil
+	}
+
+	c := *value
+
+	return &c
+}